requirements.txt
crewai
langchain-openai
pandas
openpyxl
prophet
scikit-learn
pyyaml
streamlit
plotly